+++ CavrnusSdk/Runtime/Scripts/Core/CavrnusSpaceUserHelpers.cs
@@ -11,17 +11,25 @@ namespace CavrnusCore
 		internal static IDisposable BindSpaceUsers(CavrnusSpaceConnection spaceConn, Action<CavrnusUser> userAdded, Action<CavrnusUser> userRemoved)
 		{
 			CavrnusUser lcu = null;
-			IHook lcuBind = null;
+			IDisposable lcuBind = null;
 			IDisposable mapBind = null;
 			NotifyListMapper<ISessionCommunicationRemoteUser, CavrnusUser> mapper = null;
 			
 			var spaceBind = spaceConn.CurrentSpaceConnection.Bind(sc => {
+				if (sc == null)
+					return;
+				
+				if (lcu != null) {
+					userRemoved?.Invoke(lcu);
+					lcu = null;
+				}
+				
 				lcuBind = sc.RoomSystem.Comm.LocalCommUser.Bind(lu => {
 					if (lcu != null) {
 						userRemoved?.Invoke(lcu);
 						lcu = null;
 					}
-
+				
 					if (lu != null) {
 						lcu = new CavrnusUser(lu, spaceConn);
 						userAdded?.Invoke(lcu);
@@ -36,7 +44,7 @@ namespace CavrnusCore
 				mapBind = mapper.Result.BindAll(userAdded, userRemoved);
 			});
 			
-			return lcuBind.AlsoDispose(mapper).AlsoDispose(mapBind).AlsoDispose(spaceBind);
+			return mapper.AlsoDispose(mapBind).AlsoDispose(spaceBind).AlsoDispose(lcuBind);
 		}
 	}
 }